Changed to versioning used and included an ApiController
