+++ app/controllers/api/v1/base_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V1
+    class BaseController < ApplicationController
+    end
+  end
+end
