--- app/controllers/v1/base_controller.rb
@@ -1,5 +1,0 @@
-module V1
-  class BaseController < ApplicationController
-    before_action :authenticate_user!
-  end
-end
