--- app/presenters/v1/base_presenter.rb
@@ -1,4 +1,0 @@
-module V1
-  class BasePresenter
-  end
-end
