+++ app/serializers/book_serializer.rb
@@ -1,4 +1,9 @@
 class BookSerializer < ActiveModel::Serializer
   attributes :title, :author, :genre, :publisher, :year, :image,
              :created_at, :updated_at
+  def image
+    image = Hash.new
+    image[:url] = object.image
+    image
+  end
 end
